miyakojima-web/requirements_sheets.txt
# Google Sheets Integration Requirements
# Core dependencies for miyakojima POI synchronization system

# Google Sheets API client
gspread>=5.7.0

# Google authentication
google-auth>=2.16.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0

# Environment configuration
python-dotenv>=0.19.0

# HTTP requests (dependency of gspread)
requests>=2.28.0

# Optional but recommended for better JSON handling
ujson>=5.0.0

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0